Move TOML tests into module's existing test file
Per test/CLAUDE.md: test/regression/issue/ is only for true regressions
(worked in a previous release, then broke). #28680 is a TOML
spec-compliance gap that was never correct, so the tests belong in
test/js/bun/resolve/toml/toml.test.js.
